Make wheel outlines aesthetic
Stacking two filled circles makes odd outlines
for most regular values.

But we don't really need both to be filled! The
shown outline can be regular and the inner one
can be filled. There's technically some missed
pixels but they don't seem to be noticeable in
a modicum of practice.